refactor(api): move update user password usecase
